WebGL Launch, Took out Photon from Single Player and Local Mult.
WebGL is now fully functional!

Found out that Photon was causing some weird things to happen if two people were playing the game at the same time. Probably because we were using the Photon Network to load Single Player and Local Multiplayer Scenes for some reason.

So, I took Photon out of these modes. Had to do some tweaking for WebGL but now it works perfectly!

Probably gonna add the ability to see the song title in the End Results Screen too down the line... Or maybe even add a new beat map. Who knows???

Bottom line is is that it is now a viable Portfolio piece!!! –No download required!~